+++ test-e2e/cypress-src/integration/describe-disabled-test-enabled/b-file.js
@@ -1,5 +1,6 @@
 describe("List items", () => {
   before(() => {
+    cy.task("log", "Executing before hook");
     cy.visit("/");
   });