+++ test-e2e/cypress-src/support/index.ts
@@ -1,7 +1,8 @@
+import addContext = require("mochawesome/addContext");
 import "./cypress-fail-fast";
 
 Cypress.on("test:after:run", (test, runnable) => {
   if (test.state === "failed") {
-    cy.task("log", "Executing test:after:run event in failed test");
+    addContext({ test }, "Executed test:after:run event in failed test");
   }
 });