+++ test-e2e/test/environment-config.spec.js
@@ -4,18 +4,21 @@ runSpecsTests("When it has default configuration", {
   specs: "environment-config-only",
   specsResults: [
     {
+      logBefore: true,
       executed: 4,
       passed: 1,
       failed: 1,
       skipped: 2,
     },
     {
+      logBefore: false,
       executed: 4,
       passed: 0,
       failed: 0,
       skipped: 4,
     },
     {
+      logBefore: false,
       executed: 3,
       passed: 0,
       failed: 0,
@@ -28,18 +31,21 @@ runSpecsTests("When it is disabled using plugin environment variable", {
   specs: "environment-config-only",
   specsResults: [
     {
+      logBefore: true,
       executed: 4,
       passed: 3,
       failed: 1,
       skipped: 0,
     },
     {
+      logBefore: true,
       executed: 4,
       passed: 4,
       failed: 0,
       skipped: 0,
     },
     {
+      logBefore: true,
       executed: 3,
       passed: 2,
       failed: 1,
@@ -57,18 +63,21 @@ runSpecsTests("When it is disabled using enabled environment variable", {
   specs: "environment-config-only",
   specsResults: [
     {
+      logBefore: true,
       executed: 4,
       passed: 3,
       failed: 1,
       skipped: 0,
     },
     {
+      logBefore: true,
       executed: 4,
       passed: 4,
       failed: 0,
       skipped: 0,
     },
     {
+      logBefore: true,
       executed: 3,
       passed: 2,
       failed: 1,